Move page searcher data into page searcher spec directory.
