+++ spec/services/search/page_searcher/multiple_criterion_search/multiple_search_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 require_relative '../../page_searcher/multiple_criterion_search/multiple_search_data.rb'
-require_relative '../../page_searcher_spec_data.rb'
+require_relative '../page_searcher_spec_data.rb'
 
 describe 'Search ::' do
   describe 'PageSearcher' do
