+++ spec/services/search/page_searcher/single_criterion_search/search_by_campaign_spec.rb
@@ -1,5 +1,5 @@
 require 'rails_helper'
-require_relative '../../page_searcher_spec_data.rb'
+require_relative '../page_searcher_spec_data.rb'
 
 describe 'Search ::' do
   describe 'PageSearcher' do
